kython.klogging
- move to core
- add a proper description why it's useful
- make default level INFO
- use HPI_LOGS variable for easier log level control (https://github.com/seanbreckenridge/HPI/commit/abdc6df1eaa85a4fdf7321c83d6b17f3cc344a7d)
